abstracted registerSW code in main to hook
