+++ src/main.tsx
@@ -15,7 +15,8 @@ import FAQPage from './pages/FAQ.tsx';
 import CastDetailSkeleton from './components/loadingSkeletons/CastDetailSkeleton.tsx';
 import { PersistQueryClientProvider } from '@tanstack/react-query-persist-client';
 import { createSyncStoragePersister } from '@tanstack/query-sync-storage-persister';
-import { registerSW } from 'virtual:pwa-register';
+import { RegisterSWWrapper } from './components/helpers/RegisterSWHelper.tsx';
+
 
 const CastMemberDetail = lazy(
   () => import('./pages/detailPages/CastMemberDetail.tsx'),
@@ -52,13 +53,7 @@ const queryClient = new QueryClient({
 const persister = createSyncStoragePersister({
   storage: window.localStorage,
 });
-const [updateAvailable, setUpdateAvailable] = useState(false);
 
-const updateSW = registerSW({
-  onNeedRefresh() {
-    setUpdateAvailable(true);
-  },
-});
 const router = createBrowserRouter([
   {
     path: '/',
@@ -221,20 +216,7 @@ createRoot(document.getElementById('root')!).render(
           },
         },
       }}
-    >{updateAvailable && (
-      <div className='z-100 fixed h-10 bottom-0 left-0 right-0 bg-blue-600 text-white p-4 flex justify-between items-center'>
-        <span>A new update is available!</span>
-        <button
-          className='bg-white text-blue-800 text-sm h-6 p-3 rounded flex items-center justify-center'
-          onClick={() => {
-            updateSW();
-            setUpdateAvailable(false);
-          }}
-        >
-          Reload
-        </button>
-      </div>
-    )}
+    ><RegisterSWWrapper/>
       <RouterProvider router={router} />
     </PersistQueryClientProvider>
   </StrictMode>,
