+++ src/pages/detailPages/ItemDetail.tsx
@@ -129,6 +129,7 @@ const ItemDetail = () => {
                         : 'opacity-0'
                     }`}
                     onLoad={() => setHighResPosterLoaded(true)}
+                    loading='lazy'
                     onError={(e) =>
                       ((e.target as HTMLImageElement).src =
                         '/no_poster_available.svg')
@@ -184,6 +185,7 @@ const ItemDetail = () => {
                             title={`${item.networks?.[0]?.name}`}
                             alt={`${item.networks?.[0]?.name}'s official logo`}
                             onLoad={() => setLogoLoaded(true)}
+                            loading='lazy'
                             onError={(e) =>
                               (
                                 (e.target as HTMLElement)
@@ -207,6 +209,7 @@ const ItemDetail = () => {
                             alt={`${item.production_companies?.[0]?.name}'s official logo`}
                             onLoad={() => setLogoLoaded(true)}
                             title={`${item.production_companies?.[0]?.name}`}
+                            loading='lazy'
                             onError={(e) =>
                               (
                                 (e.target as HTMLElement)
